Lots o changes -  starting implementation of JQUERY
moving along decently… have been working out some of the bugs of
changing it over to jquery. In the wimba Rick said this project was
mostly a find and replace. I have had to change quite a few things
because of the $ variable. You can see in this comment where i had to
do a "find & replace" with just that.